fix: test new login
